refactor: Move PSD class to dedicated document module
Move the PSD class from `psd_tools.psd.__init__.py` to a dedicated
`psd_tools.psd.document` module, following Python best practices of
keeping package __init__ files for exports only.

Changes:
- Created src/psd_tools/psd/document.py with PSD class implementation
- Updated src/psd_tools/psd/__init__.py to re-export PSD from document
- Updated internal imports to use psd_tools.psd.document
- Updated CLAUDE.md documentation

This is a non-breaking change - both import styles work:
- from psd_tools.psd import PSD (public API, recommended)
- from psd_tools.psd.document import PSD (internal use)

All tests pass (454 passing, 0 failures).
